Alternate method for new Reddit
As noted in 9251a75, the new Reddit design seems to be using randomised class names for everything on the page.

I can't really see a reason for this, I'm not sure whether it's deliberate obfuscation (probably not), or something to do with the framework behind it. Whatever it is, it makes stuff like this extension slightly more difficult as it's not possible to easily pick out the links that you're looking for.

The new method above simply searches all links on the page, except for those to Reddit itself. This doesn't provide whitelisting for link titles, which is something I'd like and am looking in to. Link titles should always be to Reddit itself, but it's a bit disconcerting to be clicking on something when it's not whitelisted.